[.NET] Make connector package target the strongly signed SDK packages (#661)
This is a bit of a work around for managing packages in both nuget.org
and our preview nuget feed.

The plan is for us to have connector package 0.10.1 still live in the
preview feed, but have it take a dependency on the strongly signed (and
publicly released) SDK nuget packages. Previously, the connector package
depended on _unsigned_ 0.10.0 version packages, and those aren't
published anywhere.

Also bump connector package version to 0.10.1