[Rust][Services] Fixing testing bug on Observation drop and then re-observe, same as #682 (#683)
Recreating #682 due to token issue in the CI pipeline

# Context

With recent changes added in #657, dropping an observation no longer
performs an unobserve operation. One of the state store network tests
was failing because it drops an observation and expects a subsequent
observe operation to succeed.

# Changes

The state store network test now checks for an error on an observe
operation after dropping an observation.