[Rust][Protocol] Restructure response parsing for RPC Invoker (#702)
Reimplemented and restructured response parsing for
`rpc_command::Invoker`. There is now structural separation internally
between errors reported from the Executor, and errors in parsing
responses.

This will enable error split and chunking, which were previously
unviable due to code complexity. The use of conversion traits allows for
a more modular and scalable approach, as does separating the cases of
parsing failure from successful parsing of a reported remote failure.

Note that there are a few memory inefficiencies (labelled in comments)
that will be addressed with error, split and/or chunking
implementations.